Organizando repositorio

+++ Haryel/functions.h
@@ -7,13 +7,13 @@ void doing(int* array, int size){
     getchar();
 }
 
-void showDetalhes(int* detalhes, int* array, int size, float* tempo){
+void showDetalhes(unsigned long int* detalhes, int* array, int size, float* tempo){
 
     system("clear");
     
     showArray(array, size);
-    printf("Trocas: %d\n", detalhes[1]);
-    printf("Comparacoes: %d\n", detalhes[0]);
+    printf("Trocas: %lu\n", detalhes[1]);
+    printf("Comparacoes: %lu\n", detalhes[0]);
     printf("Tempo: %f", (tempo[1] - tempo[0])/CLOCKS_PER_SEC);
     printf("\nAperte a tecla [ENTER] para continuar . . .\n");
